pyproject.toml
[project]
name = "hybrid-book-recommender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.11.0",
    "chromadb>=1.3.5",
    "dvc>=3.64.1",
    "ensure>=1.0.4",
    "google-generativeai>=0.8.5",
    "ipykernel>=7.1.0",
    "isort>=7.0.0",
    "langchain>=1.1.2",
    "matplotlib>=3.10.7",
    "mlflow>=3.7.0",
    "notebook>=7.5.0",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pytest>=9.0.1",
    "python-box>=7.3.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "streamlit>=1.52.0",
    "tqdm>=4.67.1",
    "types-pyyaml>=6.0.12.20250915",
]
